display: oplus: Automatically set AoD high/low mode based on bl_level
OPlus likely has something in userspace that would automatically decide
whether to use AoD high or low (brightness) modes for LP1 and write it
to /dev/oplus_display. We don't have that, so we're having the kernel
determine which brightness mode to use based on what the origianl
bl_level of the panel was before entering LP1 mode.

Change-Id: I85f29068377a869d0442a2f9d07d35773aefcb05