dts : Remove `display ramdump memory region` Part-2
This region is the same as the early splash framebuffer initialized by
XBL. msm_drm uses it for ramdump when available, but that's a debugging
feature we don't need. It also wastes memory after early boot because
msm_drm will keep it reserved rather than freeing the framebuffer memory
once it has taken control of the display hardware.

Remove the display ramdump region to save 8 MiB of memory and silence
the following error on boot:

[    0.000000] OF: reserved mem: OVERLAP DETECTED!\x0adisp_rdump_region (0x00000000a0000000--0x00000000a2300000) overlaps with cont_splash_region (0x00000000a0000000--0x00000000a2300000)

Credits @kdrag0n